==> 03 Scripts/01 Data Wrangling.ipynb <==
# 1. Import Libraries & Data

# import libraries
import numpy as np
import pandas as pd
import os
import datetime

# import data
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data'
delta_df = pd.read_csv(os.path.join(path, 'Original Data', 'Delta Departures Original.csv'))
american_df = pd.read_csv(os.path.join(path, 'Original Data', 'American Departures Original.csv'))
united_df = pd.read_csv(os.path.join(path, 'Original Data', 'United Departures Original.csv'))
southwest_df = pd.read_csv(os.path.join(path, 'Original Data', 'Southwest Departures Original.csv'))
airport_codes_df = pd.read_csv(os.path.join(path, 'Original Data', 'Airport Codes Original.csv'))



# 2. Drop Columns (Data Wrangling)

## airport_codes_df

# returns first five rows of airport_codes_df
airport_codes_df.head()

# checks shape of dataframe
airport_codes_df.shape

# returns list of variables for airport_codes_df
airport_codes_df.info()

# drops columns from airport_codes_df
airport_codes_df = airport_codes_df.drop(columns = {'WorldAreaCode'})

## delta_df

# returns first five rows of delta_df
delta_df.head()

# checks shape of dataframe
delta_df.shape

# returns list of variables for delta_df
delta_df.info()

# drops columns
delta_df = delta_df.drop(columns = {'Wheels-off time'})

## american_df

# returns first five rows of american_df
american_df.head()

# checks shape of dataframe
american_df.shape

# returns list of variables for alaska_df
american_df.info()

# drops columns
american_df = american_df.drop(columns = {'Wheels-off time'})

## united_df

# returns first five rows of united_df
united_df.head()

# checks shape of dataframe
united_df.shape

# returns list of variables for southwest_df
united_df.info()

# drops columns
united_df = united_df.drop(columns = {'Wheels-off time'})

## southwest_df

# returns first five rows of southwest_df
southwest_df.head()

# checks shape of dataframe
southwest_df.shape

# returns list of variables for southwest_df
southwest_df.info()

# drops columns
southwest_df = southwest_df.drop(columns = {'Wheels-off time'})

# 3. Renaming Columns (Data Wrangling)


## airport_codes_df

# renames airport_codes_df columns
airport_codes_df.rename(columns = {'AirportCode': 'Airport Code', 'CityName': 'Full Location'}, inplace = True)

# checks for renamed columns
airport_codes_df.head()

## delta_df

# renames delta_df columns
delta_df.rename(columns = {'Scheduled departure time':'Scheduled Departure Time', 'Actual departure time': 'Actual Departure Time', 'Departure delay (Minutes)': 'Departure Delay (Minutes)', 'Scheduled elapsed time (Minutes)': 'Scheduled Elapsed Time (Minutes)', 'Actual elapsed time (Minutes)': 'Actual Elapsed Time (Minutes)', 'Taxi-Out time (Minutes)': 'Taxi-Out Time (Minutes)', 'Carrier Code':'Carrier'}, inplace = True)

# checks renamed columns for delta_df
delta_df.head()

## american_df

# renames american_df columns
american_df.rename(columns = {'Scheduled departure time':'Scheduled Departure Time', 'Actual departure time': 'Actual Departure Time', 'Departure delay (Minutes)': 'Departure Delay (Minutes)', 'Scheduled elapsed time (Minutes)': 'Scheduled Elapsed Time (Minutes)', 'Actual elapsed time (Minutes)': 'Actual Elapsed Time (Minutes)', 'Taxi-Out time (Minutes)': 'Taxi-Out Time (Minutes)', 'Carrier Code':'Carrier'}, inplace = True)

# checks renamed columns for american_df
american_df.head()

## united_df

# renames united_df columns
united_df.rename(columns = {'Scheduled departure time':'Scheduled Departure Time', 'Actual departure time': 'Actual Departure Time', 'Departure delay (Minutes)': 'Departure Delay (Minutes)', 'Scheduled elapsed time (Minutes)': 'Scheduled Elapsed Time (Minutes)', 'Actual elapsed time (Minutes)': 'Actual Elapsed Time (Minutes)', 'Taxi-Out time (Minutes)': 'Taxi-Out Time (Minutes)', 'Carrier Code':'Carrier'}, inplace = True)

# checks renamed columns for united_df
united_df.head()

## southwest_df

# renames alaska_df columns
southwest_df.rename(columns = {'Scheduled departure time':'Scheduled Departure Time', 'Actual departure time': 'Actual Departure Time', 'Departure delay (Minutes)': 'Departure Delay (Minutes)', 'Scheduled elapsed time (Minutes)': 'Scheduled Elapsed Time (Minutes)', 'Actual elapsed time (Minutes)': 'Actual Elapsed Time (Minutes)', 'Taxi-Out time (Minutes)': 'Taxi-Out Time (Minutes)', 'Carrier Code':'Carrier'}, inplace = True)

# checks renamed columns for southwest_df
southwest_df.head()

# 4. Checking for Mixed & Data Types (Data Types)

## airport_codes_df

# check for mixed types in the airport_codes_df
for cols in airport_codes_df.columns.tolist():
    mixed_type = (airport_codes_df[[cols]].applymap(type) != airport_codes_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(airport_codes_df[mixed_type])>0:
        print(cols)
# no mixed type columnes

## delta_df

# check for mixed types in the delta_df
for cols in delta_df.columns.tolist():
    mixed_type = (delta_df[[cols]].applymap(type) != delta_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(delta_df[mixed_type])>0:
        print(cols)

# converts mixed type columns to strings and date times
delta_df['Carrier'] = delta_df['Carrier'].astype(str)
delta_df['Date (MM/DD/YYYY)'] = delta_df['Date (MM/DD/YYYY)'].astype(str)
delta_df['Tail Number'] = delta_df['Tail Number'].astype(str)
delta_df['Destination Airport'] = delta_df['Destination Airport'].astype(str)
delta_df['Actual Departure Time'] = pd.to_datetime(delta_df['Actual Departure Time'], format = '%H:%M').dt.time
delta_df['Scheduled Departure Time'] = pd.to_datetime(delta_df['Scheduled Departure Time'], format = '%H:%M').dt.time


# check for mixed types in the delta_df
for cols in delta_df.columns.tolist():
    mixed_type = (delta_df[[cols]].applymap(type) != delta_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(delta_df[mixed_type])>0:
        print(cols)
# Scheduled Departure Times & Actual Departure times have na values in them, resulting in mixed values.These will be taken care of in next script.


## american_df

# check for mixed types in the american_df
for cols in american_df.columns.tolist():
    mixed_type = (american_df[[cols]].applymap(type) != american_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(american_df[mixed_type])>0:
        print(cols)

# converts mixed type columns to strings and date times
american_df['Carrier'] = american_df['Carrier'].astype(str)
american_df['Date (MM/DD/YYYY)'] = american_df['Date (MM/DD/YYYY)'].astype(str)
american_df['Tail Number'] = american_df['Tail Number'].astype(str)
american_df['Destination Airport'] = american_df['Destination Airport'].astype(str)
american_df['Actual Departure Time'] = pd.to_datetime(american_df['Actual Departure Time'], format = '%H:%M').dt.time
american_df['Scheduled Departure Time'] = pd.to_datetime(american_df['Scheduled Departure Time'], format = '%H:%M').dt.time

# check for mixed types in the american_df
for cols in american_df.columns.tolist():
    mixed_type = (american_df[[cols]].applymap(type) != american_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(american_df[mixed_type])>0:
        print(cols)
# Scheduled Departure Times & Actual Departure times have na values in them, resulting in mixed values.These will be taken care of in next script.

## united_df

# check for mixed types in the united_df
for cols in united_df.columns.tolist():
    mixed_type = (united_df[[cols]].applymap(type) != united_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(united_df[mixed_type])>0:
        print(cols)

# converts mixed type columns to strings and date times
united_df['Carrier'] = united_df['Carrier'].astype(str)
united_df['Date (MM/DD/YYYY)'] = united_df['Date (MM/DD/YYYY)'].astype(str)
united_df['Tail Number'] = united_df['Tail Number'].astype(str)
united_df['Destination Airport'] = united_df['Destination Airport'].astype(str)
united_df['Actual Departure Time'] = pd.to_datetime(united_df['Actual Departure Time'], format = '%H:%M').dt.time
united_df['Scheduled Departure Time'] = pd.to_datetime(united_df['Scheduled Departure Time'], format = '%H:%M').dt.time


# check for mixed types in the united_df
for cols in united_df.columns.tolist():
    mixed_type = (united_df[[cols]].applymap(type) != united_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(united_df[mixed_type])>0:
        print(cols)
# Scheduled Departure Times & Actual Departure times have na values in them, resulting in mixed values.These will be taken care of in next script.

## southwest_df

# check for mixed types in the southwest_df
for cols in southwest_df.columns.tolist():
    mixed_type = (southwest_df[[cols]].applymap(type) != southwest_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(southwest_df[mixed_type])>0:
        print(cols)

# converts mixed type columns to strings and date times
southwest_df['Carrier'] = southwest_df['Carrier'].astype(str)
southwest_df['Date (MM/DD/YYYY)'] = southwest_df['Date (MM/DD/YYYY)'].astype(str)
southwest_df['Tail Number'] = southwest_df['Tail Number'].astype(str)
southwest_df['Destination Airport'] = southwest_df['Destination Airport'].astype(str)
southwest_df['Actual Departure Time'] = pd.to_datetime(southwest_df['Actual Departure Time'], format = '%H:%M').dt.time
southwest_df['Scheduled Departure Time'] = pd.to_datetime(southwest_df['Scheduled Departure Time'], format = '%H:%M').dt.time

# check for mixed types in the southwest_df
for cols in southwest_df.columns.tolist():
    mixed_type = (southwest_df[[cols]].applymap(type) != southwest_df
[[cols]].iloc[0].apply(type)).any(axis = 1)
    if len(southwest_df[mixed_type])>0:
        print(cols)
# Scheduled Departure Times & Actual Departure times have na values in them, resulting in mixed values.These will be taken care of in next script.



# 5. Change Data Types

# Changes data types of flight number for four airlines to a string.
delta_df['Flight Number'] = delta_df['Flight Number'].astype(str)
american_df['Flight Number'] = american_df['Flight Number'].astype(str)
united_df['Flight Number'] = united_df['Flight Number'].astype(str)
southwest_df['Flight Number'] = southwest_df['Flight Number'].astype(str)




# 6. Transpose Data & Subset
## No data needs to be transposed

# 7. Export Wrangled Dataframes


# exports all dfs to prepared data folder
delta_df.to_csv(os.path.join(path, 'Prepared Data', 'delta_df_wrangled.csv'))
american_df.to_csv(os.path.join(path, 'Prepared Data', 'american_df_wrangled.csv'))
united_df.to_csv(os.path.join(path, 'Prepared Data', 'united_df_wrangled.csv'))
southwest_df.to_csv(os.path.join(path, 'Prepared Data', 'southwest_df_wrangled.csv'))
airport_codes_df.to_csv(os.path.join(path, 'Prepared Data', 'airport_codes_df_wrangled.csv'))
