==> 03 Scripts/02 Clean Data.ipynb <==
# 1. Import Libraries & Data

# import libraries
import numpy as np
import pandas as pd
import os
import datetime

# import data
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data'
delta_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'delta_df_wrangled.csv'), index_col = [0])
american_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'american_df_wrangled.csv'), index_col = [0])
united_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'united_df_wrangled.csv'), index_col = [0])
southwest_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'southwest_df_wrangled.csv'), index_col = [0])
airport_codes_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'airport_codes_df_wrangled.csv'), index_col = [0])



# 2. Convert mixed datatypes

# converts mixed type columns to str 
delta_df['Carrier'] = delta_df['Carrier'].astype(str)
delta_df['Date (MM/DD/YYYY)'] = delta_df['Date (MM/DD/YYYY)'].astype(str)
delta_df['Tail Number'] = delta_df['Tail Number'].astype(str)
delta_df['Destination Airport'] = delta_df['Destination Airport'].astype(str)
american_df['Carrier'] = american_df['Carrier'].astype(str)
american_df['Date (MM/DD/YYYY)'] = american_df['Date (MM/DD/YYYY)'].astype(str)
american_df['Tail Number'] = american_df['Tail Number'].astype(str)
american_df['Destination Airport'] = american_df['Destination Airport'].astype(str)
united_df['Carrier'] = united_df['Carrier'].astype(str)
united_df['Date (MM/DD/YYYY)'] = united_df['Date (MM/DD/YYYY)'].astype(str)
united_df['Tail Number'] = united_df['Tail Number'].astype(str)
united_df['Destination Airport'] = united_df['Destination Airport'].astype(str)
southwest_df['Carrier'] = southwest_df['Carrier'].astype(str)
southwest_df['Date (MM/DD/YYYY)'] = southwest_df['Date (MM/DD/YYYY)']
southwest_df['Tail Number'] = southwest_df['Tail Number'].astype(str)
southwest_df['Destination Airport'] = southwest_df['Destination Airport'].astype(str)
delta_df['Flight Number'] = delta_df['Flight Number'].astype(str)
american_df['Flight Number'] = american_df['Flight Number'].astype(str)
united_df['Flight Number'] = united_df['Flight Number'].astype(str)
southwest_df['Flight Number'] = southwest_df['Flight Number'].astype(str)

# converts mixed type columns to time
delta_df['Actual Departure Time'] = pd.to_datetime(delta_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
delta_df['Scheduled Departure Time'] = pd.to_datetime(delta_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
united_df['Actual Departure Time'] = pd.to_datetime(united_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
united_df['Scheduled Departure Time'] = pd.to_datetime(united_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
american_df['Actual Departure Time'] = pd.to_datetime(american_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
american_df['Scheduled Departure Time'] = pd.to_datetime(american_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
southwest_df['Actual Departure Time'] = pd.to_datetime(southwest_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
southwest_df['Scheduled Departure Time'] = pd.to_datetime(southwest_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time

# 3. Check for missing values

## airport_codes_df

#checks for missing values in airport_codes_df
airport_codes_df.isnull().sum()
# no null values

## delta_df 

#checks for missing values in delta_df
delta_df.isnull().sum()
# missing values make up less than 5% of dataset. Choosing to drop missing values.

# drops missing values from delta_df
delta_df = delta_df.dropna()

delta_df.shape

## american_df 

#checks for missing values in american_df
american_df.isnull().sum()
# missing values make up less than 5% of dataset. Choosing to drop missing values.

# drops missing values from american_df 
american_df = american_df.dropna()

## united_df

#checks for missing values in united_df
united_df.isnull().sum()
# missing values make up less than 5% of dataset. Choosing to drop missing values.

# drops missing values from united_df 
united_df = united_df.dropna()

## southwest_df

#checks for missing values in southwest_df
southwest_df.isnull().sum()
# missing values make up less than 5% of dataset. Choosing to drop missing values.

# drops missing values from southwest_df
southwest_df = southwest_df.dropna()

# 4. Duplicates

## airport_codes_df 

# checks for full duplicates in airport_codes_df
airport_codes_df[airport_codes_df.duplicated()]
# no duplicates

## delta_df

# checks for full duplicates in delta_df 
delta_df[delta_df.duplicated()]
# no duplicates

## american_df

# checks for full duplicates in american_df 
american_df[american_df.duplicated()]
# no duplicates

## united_df

# checks for full duplicates in united_df 
united_df[united_df.duplicated()]
#no duplicates

## southwest_df

# checks for full duplicates in southwest_df 
southwest_df[southwest_df.duplicated()]
# no duplicates

# 5. Accuracy Checks

## Descriptive stats (accuracy check)

### airport_codes_df

# descriptive stats for airport_codes_df
airport_codes_df.describe()
## no need to pull value counts at this point - too many unique

### delta_df (numerical data)

delta_df.info()

# descriptive stats for delta_df
delta_df.describe()
# negative numbers could mean the flight left early? 

### delta_df (categorical data)

# returns counts for delta_df['Carrier Code']
delta_df['Carrier'].value_counts()

# replaces abbreviated airline name with full airline name
delta_df['Carrier'] = delta_df['Carrier'].replace('DL', 'Delta')

# Checks output of airline name
delta_df.head()

# returns counts for delta_df['Tail Number']
delta_df['Tail Number'].value_counts()

# nan values for delta_df['Tail Number']
nan_delta = delta_df[delta_df['Tail Number'] == 'nan']

nan_delta.shape

delta_df['Flight Number'].value_counts()

# returns value counts for delta_df['Destination Airport']
delta_df['Destination Airport'].value_counts()


### american_df (numerical data)

# descriptive stats for american_df
american_df.describe()
# negative numbers could mean the flight left early? 

### american_df (categorical data)

# returns counts for american_df['Carrier']
american_df['Carrier'].value_counts()

# replaces abbreviated airline name with full airline name
american_df['Carrier'] = american_df['Carrier'].replace('AA', 'American')

# Checks output of airline name
american_df['Carrier'].head()

# returns counts for american_df['Tail Number']
american_df['Tail Number'].value_counts()

# nan values for american_df['Tail Number']
nan_american = american_df[american_df['Tail Number'] == 'nan']

# prints nan values count
nan_american.shape

# returns counts for american_df['Flight Number']
american_df['Flight Number'].value_counts()

# returns value counts for american_df['Destination Airport']
american_df['Destination Airport'].value_counts()

### united_df (numerical data)


# descriptive stats for united_df
united_df.describe()
# negative numbers could mean the flight left early? 

### united_df (categorical data)

# returns counts for united_df['Carrier']
united_df['Carrier'].value_counts()

# replaces abbreviated airline name with full airline name
united_df['Carrier'] = united_df['Carrier'].replace('UA', 'United')

# Checks output of airline name
united_df['Carrier'].head()

# returns counts for united_df['Tail Number'] leave in nan values
united_df['Tail Number'].value_counts()

# nan values for united_df['Tail Number']
nan_united = united_df[united_df['Tail Number'] == 'nan']

# prints nan values count
nan_united.shape

# returns counts for united_df['Tail Number']
united_df['Flight Number'].value_counts()

# returns counts for united_df['Tail Number']
united_df['Destination Airport'].value_counts()

### southwest_df (numerical data)

# descriptive stats for southwest_df
southwest_df.describe()
# negative numbers could mean the flight left early? 

### southwest_df (categorical data)

# returns counts for southwest_df['Carrier']
southwest_df['Carrier'].value_counts()

# replaces abbreviated airline name with full airline name
southwest_df['Carrier'] = southwest_df['Carrier'].replace('WN', 'Southwest')

# checks output of full airline name
southwest_df['Carrier'].head()

# returns counts for southwest_df['Tail Number']
southwest_df['Tail Number'].value_counts()

# nan values for southwest_df['Tail Number']
nan_southwest = southwest_df[southwest_df['Tail Number'] == 'nan']

# prints nan values count
nan_southwest.shape

# returns counts for southwest_df['Flight Number']
southwest_df['Flight Number'].value_counts()

# returns counts for southwest_df['Destination Airport']
southwest_df['Destination Airport'].value_counts()

# Note, times, dates, & locations have not been investigated yet, as transformations are needed on their values

# 5. Check Shapes of cleaned data frames

# checks shape of delta_df
delta_df.shape

# checks shape of american_df
american_df.shape

# checks shape of united_df
united_df.shape

# checks shape of southwest_df
southwest_df.shape

# checks shape of airport_codes_df
airport_codes_df.shape

# 6. Export Data

# exports all dfs to prepared data folder
delta_df.to_csv(os.path.join(path, 'Prepared Data', 'delta_df_cleaned.csv'))
american_df.to_csv(os.path.join(path, 'Prepared Data', 'american_df_cleaned.csv'))
united_df.to_csv(os.path.join(path, 'Prepared Data', 'united_df_cleaned.csv'))
southwest_df.to_csv(os.path.join(path, 'Prepared Data', 'southwest_df_cleaned.csv'))
airport_codes_df.to_csv(os.path.join(path, 'Prepared Data', 'airport_codes_df_cleaned.csv'))
