==> 03 Scripts/03 Merge Dataframes.ipynb <==
# 1. Import Libraries & Data

# import libraries
import numpy as np
import pandas as pd
import os
import datetime

# import data
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data'
delta_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'delta_df_cleaned.csv'), index_col = [0])
american_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'american_df_cleaned.csv'), index_col = [0])
united_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'united_df_cleaned.csv'), index_col = [0])
southwest_df =  pd.read_csv(os.path.join(path, 'Prepared Data', 'southwest_df_cleaned.csv'), index_col = [0])
airport_codes_df= pd.read_csv(os.path.join(path, 'Prepared Data', 'airport_codes_df_cleaned.csv'), index_col = [0])



# converts mixed type columns to str
delta_df['Carrier'] = delta_df['Carrier'].astype(str)
delta_df['Date (MM/DD/YYYY)'] = delta_df['Date (MM/DD/YYYY)'].astype(str)
delta_df['Tail Number'] = delta_df['Tail Number'].astype(str)
delta_df['Destination Airport'] = delta_df['Destination Airport'].astype(str)
american_df['Carrier'] = american_df['Carrier'].astype(str)
american_df['Date (MM/DD/YYYY)'] = american_df['Date (MM/DD/YYYY)'].astype(str)
american_df['Tail Number'] = american_df['Tail Number'].astype(str)
american_df['Destination Airport'] = american_df['Destination Airport'].astype(str)
united_df['Carrier'] = united_df['Carrier'].astype(str)
united_df['Date (MM/DD/YYYY)'] = united_df['Date (MM/DD/YYYY)'].astype(str)
united_df['Tail Number'] = united_df['Tail Number'].astype(str)
united_df['Destination Airport'] = united_df['Destination Airport'].astype(str)
southwest_df['Carrier'] = southwest_df['Carrier'].astype(str)
southwest_df['Date (MM/DD/YYYY)'] = southwest_df['Date (MM/DD/YYYY)']
southwest_df['Tail Number'] = southwest_df['Tail Number'].astype(str)
southwest_df['Destination Airport'] = southwest_df['Destination Airport'].astype(str)
delta_df['Flight Number'] = delta_df['Flight Number'].astype(str)
american_df['Flight Number'] = american_df['Flight Number'].astype(str)
united_df['Flight Number'] = united_df['Flight Number'].astype(str)
southwest_df['Flight Number'] = southwest_df['Flight Number'].astype(str)

# converts mixed type columns to time
delta_df['Actual Departure Time'] = pd.to_datetime(delta_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
delta_df['Scheduled Departure Time'] = pd.to_datetime(delta_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
united_df['Actual Departure Time'] = pd.to_datetime(united_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
united_df['Scheduled Departure Time'] = pd.to_datetime(united_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
american_df['Actual Departure Time'] = pd.to_datetime(american_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
american_df['Scheduled Departure Time'] = pd.to_datetime(american_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time
southwest_df['Actual Departure Time'] = pd.to_datetime(southwest_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
southwest_df['Scheduled Departure Time'] = pd.to_datetime(southwest_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time


# 2. Return list of variables

# returns variable list for delta_df
delta_df.info()

# returns variable list for american_df
american_df.info()

# returns variable list for southwest_df
southwest_df.info()

# returns variable list for united_df
united_df.info()

# returns variable list for airport_codes_df
airport_codes_df.info()

# 3. Append flight info dataframes

# appends delta_df & american_df
airlines_df = delta_df.append(american_df)

# appends airlines df & southwest_df
airlines_df = airlines_df.append(southwest_df)

# appends airlines_df & united_df
airlines_df = airlines_df.append(united_df)

# checks shape of airlines_df
airlines_df.shape

# 4. Merge airlines_df and airport_codes_df

# renames airport_codes_df{'Airport Code'} in preparation for merge
airport_codes_df.rename(columns = {'Airport Code': 'Destination Airport'}, inplace = True)

# merges airlines_df and airport_codes_df
airlines_final_df = airlines_df.merge(airport_codes_df, on = 'Destination Airport', indicator = True)

# checks for result of the merge
airlines_final_df.head()

# checks for counts of indicator 
airlines_final_df['_merge'].value_counts()


# drops indicator column
airlines_final_df = airlines_final_df.drop(columns = ['_merge'])

# checks shape of airlines_final_df
airlines_final_df.shape

# 5. Export Dataframe

# exports dataframe
airlines_final_df.to_csv(os.path.join(path, 'Prepared Data', 'airlines_final_df.csv'))
