==> 03 Scripts/04. Derive New Variables.ipynb <==
# Import Libraries

# import libraries
import numpy as np
import pandas as pd
import os
import datetime

# import dataframe
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data'
airlines_final_df = pd.read_csv(os.path.join(path, 'Prepared Data', 'airlines_final_df.csv'), index_col = [0])

# Convert Mixed Type Columns

# converts datatpes
airlines_final_df['Carrier'] = airlines_final_df['Carrier'].astype(str)
airlines_final_df['Date (MM/DD/YYYY)'] = airlines_final_df['Date (MM/DD/YYYY)'].astype(str)
airlines_final_df['Tail Number'] = airlines_final_df['Tail Number'].astype(str)
airlines_final_df['Destination Airport'] = airlines_final_df['Destination Airport'].astype(str)
airlines_final_df['Flight Number'] = airlines_final_df['Flight Number'].astype(str)
airlines_final_df['Actual Departure Time'] = pd.to_datetime(airlines_final_df['Actual Departure Time'], format = '%H:%M:%S').dt.time
airlines_final_df['Scheduled Departure Time'] = pd.to_datetime(airlines_final_df['Scheduled Departure Time'], format = '%H:%M:%S').dt.time

# Create new variables

# returns first five rows of airlines_final_df
airlines_final_df.head()

## create new variables from Full Location column

# splits Full Location into four columns
airlines_final_df[['City', 'State', 'Country', 'Placeholder']] = airlines_final_df['Full Location'].str.split(', ', expand = True)

# drops extra column
airlines_final_df = airlines_final_df.drop(columns = {'Placeholder', 'Full Location'})

airlines_final_df.head()

## creates new columns from Date (MM/DD/YYYY) column

#  Splits flight date into day, month, and year
airlines_final_df[['Flight Month', 'Flight Day', 'Flight Year']] =airlines_final_df['Date (MM/DD/YYYY)'].str.split('/', expand = True)

# checks output after split and dropped columns
airlines_final_df.head()

## create new variable for known delays

## create new variable for the total delay time with delays with known causes
airlines_final_df['Known Total Delay (Minutes)'] = airlines_final_df['Delay Weather (Minutes)']+airlines_final_df['Delay Carrier (Minutes)']+ airlines_final_df['Delay National Aviation System (Minutes)']+airlines_final_df['Delay Security (Minutes)']+airlines_final_df['Delay Late Aircraft Arrival (Minutes)']

## checks first five rows of new variable
airlines_final_df['Known Total Delay (Minutes)'].head()

## creates variable for descrepancy

## creates variable for unaccounted delay time
airlines_final_df['Unaccounted Delay Time'] = airlines_final_df['Known Total Delay (Minutes)'] - airlines_final_df['Departure Delay (Minutes)'] 

## checks first five rows of new variable
airlines_final_df['Unaccounted Delay Time'].head()

# Change value formats in split columns

# changes capitalization for airlines_final_df['City']
airlines_final_df['City'] = airlines_final_df['City'].str.title()

# changes capitalization for airlines_final_df['State']
airlines_final_df['State'] = airlines_final_df['State'].str.title()

# changes values for airlines_final_df['Flight Month'] from numbers to month names
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('1', 'January')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('2', 'February')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('3', 'March')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('4', 'April')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('5', 'May')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('6', 'June')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('7', 'July')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('8', 'August')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('9', 'September')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('10', 'October')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('11', 'November')
airlines_final_df['Flight Month'] = airlines_final_df['Flight Month'].replace('12', 'December')

# changes airlines_final_df['Flight Year'] to long version of year
airlines_final_df['Flight Year'] = airlines_final_df['Flight Year'].replace('17', '2017')
airlines_final_df['Flight Year'] = airlines_final_df['Flight Year'].replace('18', '2018')
airlines_final_df['Flight Year'] = airlines_final_df['Flight Year'].replace('19', '2019')
airlines_final_df['Flight Year'] = airlines_final_df['Flight Year'].replace('20', '2020')
airlines_final_df['Flight Year'] = airlines_final_df['Flight Year'].replace('21', '2021')

# Check accuracy/consistency of new values

# airlines_final_df['Flight Month'] value counts
airlines_final_df['Flight Month'].value_counts()

# airlines_final_df['Flight Year'] value counts
airlines_final_df['Flight Year'].value_counts()

# airlines_final_df['Flight Day'] value counts
airlines_final_df['Flight Day'].value_counts()

# airlines_final_df['Flight City'] value counts
airlines_final_df['City'].value_counts()

# airlines_final_df['State'] value counts
airlines_final_df['State'].value_counts()
# accuracy issue: some of these are not states

# replaces places that are not states with state name
airlines_final_df['State'] = airlines_final_df['State'].replace('Oahu', 'Hawaii')
airlines_final_df['State'] = airlines_final_df['State'].replace('Maui', 'Hawaii')
airlines_final_df['State'] = airlines_final_df['State'].replace('Kauai', 'Hawaii')

# airlines_final_df['State'] value counts
airlines_final_df['State'].value_counts()

# airlines_final_df['Carrier'] value counts
airlines_final_df['Carrier'].value_counts()

# checks value counts for ['Country']
airlines_final_df['Country'].value_counts()
# accuracy issue: Hawaii is not a state
# all locations are USA, so column can be dropped.

# drops Country column
airlines_final_df = airlines_final_df.drop(columns = {'Country'})




# checks new output
airlines_final_df.head()

# checks shape 
airlines_final_df.shape

# difference in departure time (expected vs real)`

# creates variable for the Actual Departure Time & Date
airlines_final_df['Actual Departure Date & Time'] =(airlines_final_df['Date (MM/DD/YYYY)']) + ' ' +(airlines_final_df['Actual Departure Time'].astype(str))

# creates variable for the Actual Departure Time & Date
airlines_final_df['Scheduled Departure Date & Time'] =(airlines_final_df['Date (MM/DD/YYYY)']) + ' ' +(airlines_final_df['Scheduled Departure Time'].astype(str))

# converts Actual Departure Date & Time to datetime and creates a new object for it.
actual_dep_datetime = pd.to_datetime(airlines_final_df['Actual Departure Date & Time'], format = '%m/%d/%y %H:%M:%S')

# converts Scheduled Departure Date & Time to datetime and creates a new object for it.
scheduled_dep_datetime = pd.to_datetime(airlines_final_df['Scheduled Departure Date & Time'], format = '%m/%d/%y %H:%M:%S')

# creates columns for the Difference in Actual & Expected Departure Time
airlines_final_df['Difference in Actual & Expected Departure Time'] = actual_dep_datetime-scheduled_dep_datetime

# outputs the first five rows of airlines_final_df['Difference in Actual & Expected Departure Time']
airlines_final_df['Difference in Actual & Expected Departure Time'].head()

# difference in time in air (expected vs real)`

# creates variable for the Difference in Actual & Expected Air Time
airlines_final_df['Difference in Actual & Expected Air Time'] =airlines_final_df['Scheduled Elapsed Time (Minutes)'] - airlines_final_df['Actual Elapsed Time (Minutes)']

# Normalized Data

# outputs total delay in minutes for each carrier
airlines_final_df.groupby('Carrier')['Known Total Delay (Minutes)'].sum()

# normalizes delays for Southwest
airlines_final_df.loc[airlines_final_df['Carrier']=='Southwest', ['Delay Carrier (%)']]= (airlines_final_df['Delay Carrier (Minutes)']/1241658)
airlines_final_df.loc[airlines_final_df['Carrier']=='Southwest', ['Delay Weather (%)']]= (airlines_final_df['Delay Weather (Minutes)']/1241658)
airlines_final_df.loc[airlines_final_df['Carrier']=='Southwest', ['Delay National Aviation System (%)']]= (airlines_final_df['Delay National Aviation System (Minutes)']/1241658)
airlines_final_df.loc[airlines_final_df['Carrier']=='Southwest', ['Delay Security (%)']]= (airlines_final_df['Delay Security (Minutes)']/1241658)
airlines_final_df.loc[airlines_final_df['Carrier']=='Southwest', ['Delay Late Aircraft Arrival (%)']]= (airlines_final_df['Delay Late Aircraft Arrival (Minutes)']/1241658)

# normalizes delays for United
airlines_final_df.loc[airlines_final_df['Carrier']=='United', ['Delay Carrier (%)']]= (airlines_final_df['Delay Carrier (Minutes)']/427991)
airlines_final_df.loc[airlines_final_df['Carrier']=='United', ['Delay Weather (%)']]= (airlines_final_df['Delay Weather (Minutes)']/427991)
airlines_final_df.loc[airlines_final_df['Carrier']=='United', ['Delay National Aviation System (%)']]= (airlines_final_df['Delay National Aviation System (Minutes)']/427991)
airlines_final_df.loc[airlines_final_df['Carrier']=='United', ['Delay Security (%)']]= (airlines_final_df['Delay Security (Minutes)']/427991)
airlines_final_df.loc[airlines_final_df['Carrier']=='United', ['Delay Late Aircraft Arrival (%)']]= (airlines_final_df['Delay Late Aircraft Arrival (Minutes)']/427991)

# normalizes delays for American
airlines_final_df.loc[airlines_final_df['Carrier']=='American', ['Delay Carrier (%)']]= (airlines_final_df['Delay Carrier (Minutes)']/434469)
airlines_final_df.loc[airlines_final_df['Carrier']=='American', ['Delay Weather (%)']]= (airlines_final_df['Delay Weather (Minutes)']/434469)
airlines_final_df.loc[airlines_final_df['Carrier']=='American', ['Delay National Aviation System (%)']]= (airlines_final_df['Delay National Aviation System (Minutes)']/434469)
airlines_final_df.loc[airlines_final_df['Carrier']=='American', ['Delay Security (%)']]= (airlines_final_df['Delay Security (Minutes)']/434469)
airlines_final_df.loc[airlines_final_df['Carrier']=='American', ['Delay Late Aircraft Arrival (%)']]= (airlines_final_df['Delay Late Aircraft Arrival (Minutes)']/434469)

# normalizes delays for Delta
airlines_final_df.loc[airlines_final_df['Carrier']=='Delta', ['Delay Carrier (%)']]= (airlines_final_df['Delay Carrier (Minutes)']/231787)
airlines_final_df.loc[airlines_final_df['Carrier']=='Delta', ['Delay Weather (%)']]= (airlines_final_df['Delay Weather (Minutes)']/231787)
airlines_final_df.loc[airlines_final_df['Carrier']=='Delta', ['Delay National Aviation System (%)']]= (airlines_final_df['Delay National Aviation System (Minutes)']/231787)
airlines_final_df.loc[airlines_final_df['Carrier']=='Delta', ['Delay Security (%)']]= (airlines_final_df['Delay Security (Minutes)']/231787)
airlines_final_df.loc[airlines_final_df['Carrier']=='Delta', ['Delay Late Aircraft Arrival (%)']]= (airlines_final_df['Delay Late Aircraft Arrival (Minutes)']/231787)

# quick check for duplicates and null values

# Check for null values
airlines_final_df.isnull().sum()

# check for duplicates
airlines_final_df[airlines_final_df.duplicated()]
# no duplicates

# Delay?

#creates variable for minutes delayed
airlines_final_df.loc[airlines_final_df['Known Total Delay (Minutes)']>0, ['Delay?']]='Yes'
airlines_final_df.loc[airlines_final_df['Known Total Delay (Minutes)']<=0, ['Delay?']]='No'



airlines_final_df['Delay?'].value_counts()

# export dataframe

# exports dataframe
airlines_final_df.to_csv(os.path.join(path, 'Prepared Data', 'airlines_final_df_new_variables.csv'))
