==> 03 Scripts/05 Descriptive Stats for New Dataframe.ipynb <==
# Import Data & Libraries

# import libraries
import pandas as pd
import numpy as np
import os

# import data frame
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data'
airlines_final_df= pd.read_csv(os.path.join(path, 'Prepared Data', 'airlines_final_df_new_variables.csv'), index_col = [0])

# Convert Data Types

# converts data types
airlines_final_df['Carrier'] = airlines_final_df['Carrier'].astype(str)
airlines_final_df['Tail Number'] = airlines_final_df['Tail Number'].astype(str)
airlines_final_df['Destination Airport'] = airlines_final_df['Destination Airport'].astype(str)
airlines_final_df['Actual Departure Time'] = pd.to_datetime(airlines_final_df['Actual Departure Time']).dt.hour
airlines_final_df['Scheduled Departure Time'] = pd.to_datetime(airlines_final_df['Scheduled Departure Time']).dt.hour
airlines_final_df['Flight Number'] = airlines_final_df['Flight Number'].astype(str)

# Descriptive Stats (numerical variables)

# returns variable list
airlines_final_df.info()

# saves descriptive stats to dataframe
descriptive_stats_air = airlines_final_df.describe()

# Descriptive Stats (categorical variables)

# value counts for airlines_final_df['City']
airlines_final_df['City'].value_counts()

# value counts for airlines_final_df['Carrier']
airlines_final_df['Carrier'].value_counts()

# value counts for airlines_final_df['State']
airlines_final_df['State'].value_counts()

# value counts for airlines_final_df['Destination Airport']
airlines_final_df['Destination Airport'].value_counts()

# value counts for airlines_final_df['Flight Month']
airlines_final_df['Flight Month'].value_counts()

# value counts for airlines_final_df['Flight Year']
airlines_final_df['Flight Year'].value_counts()

# value counts for airlines_final_df['Flight Day']
airlines_final_df['Flight Day'].value_counts().sort_index()

# saves descriptive_stats_air to file
path2 = r'/Users/oliviamurphy/Desktop/Airlines Project/04 Analysis'
descriptive_stats_air.to_csv(os.path.join(path2, 'descriptive_stats_air.csv'))
