==> 03 Scripts/11 Additional Analysis.ipynb <==
# Import Libraries

# imports libraires
import numpy as np
import pandas as pd
import os

# Import Data

# imports data
path = r'/Users/oliviamurphy/Desktop/Airlines Project/02 Data/Prepared Data'
airlines_df = pd.read_csv(os.path.join(path, 'airlines_df_cluster.csv'))

# Stats to Tell Tableau Story

# returns columns
airlines_df.info()

# number of flights leaving SFI from 2017-2021
airlines_df.shape

# number of flights leaving SFI in 2021
airlines_df['Flight Year'].value_counts()

df_2021 = airlines_df[airlines_df['Flight Year']== 2021]

df_2021['Delay?'].value_counts()

6111/35817

# Create dataframe for ontime flights vs delayed flights

# creates subset for delayed flights
delayed_flights = airlines_df[airlines_df['Delay?']=='Yes']

# creates subset for ontime flights
ontime_flights = airlines_df[airlines_df['Delay?']=='No']

# creates df for counts of delayed  flights
delayed_flights_grouped = delayed_flight.groupby(['Carrier', 'Flight Year'])['Delay?'].count()

# checks output of delayed_flights_grouped
delayed_flights_grouped

# resets index for delayed_flights_grouped
delayed_flights_grouped = delayed_flights_grouped.reset_index()

# renames column for the count of the delayed flights
delayed_flights_grouped.rename(columns = {'Delay?': 'Delayed Flights Count'}, inplace = True)

# checks output of delayed_flights_grouped
delayed_flights_grouped

# # creates df for counts of ontime  flights
ontime_flights_grouped = ontime_flights.groupby(['Carrier', 'Flight Year'])['Delay?'].count()

# checks output of ontime_flights_grouped
ontime_flights_grouped

#resets index for ontime flights
ontime_flights_grouped = ontime_flights_grouped.reset_index()

# renames column for ontime flight counts 
ontime_flights_grouped.rename(columns = {'Delay?': 'Ontime Flights Count'}, inplace = True)

# Checks outout of ontime_flights_grouped
ontime_flights_grouped

# merges dataframes
delayed_ontime_flights = ontime_flights_grouped.merge(delayed_flights_grouped, on = ['Carrier', 'Flight Year'])

# checks output of merge
delayed_ontime_flights

# creates column for total flights
delayed_ontime_flights['Total Flights'] = delayed_ontime_flights['Ontime Flights Count']+ delayed_ontime_flights['Delayed Flights Count']

# creates column for the % of ontime flights
delayed_ontime_flights['Ontime Flights %'] = delayed_ontime_flights['Ontime Flights Count']/delayed_ontime_flights['Total Flights']

# creates column for the % of delayed flights
delayed_ontime_flights['Delayed Flights %'] = delayed_ontime_flights['Delayed Flights Count']/delayed_ontime_flights['Total Flights']

# checks output 
delayed_ontime_flights

# exports dataframe to csv
delayed_ontime_flights.to_csv(os.path.join(path, 'ontime_vs_delayed_overview.csv'))
